backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
psycopg[binary]==3.2.10
alembic==1.16.4
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic-settings==2.11.0
email-validator==2.3.0
qrcode[pil]==8.2
pypdf==5.9.0
python-docx==1.2.0
pytest==8.4.1
httpx==0.28.1
